requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==20.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP Requests
httpx==0.25.2

# System Monitoring
psutil==5.9.6

# Environment Variables
python-dotenv==1.0.0


# Utilities
python-dateutil==2.8.2
